+++ src/commands/hardis/org/purge/flow.ts
@@ -4,9 +4,9 @@ import { Messages, SfError } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import c from 'chalk';
 import columnify from 'columnify';
-import { execSfdxJson, isCI, uxLog } from '../../../../common/utils/index.js';
+import { execSfdxJson, extractRegexMatches, isCI, uxLog } from '../../../../common/utils/index.js';
 import { prompts } from '../../../../common/utils/prompts.js';
-import { bulkDeleteTooling } from '../../../../common/utils/apiUtils.js';
+import { bulkDelete, bulkDeleteTooling, bulkQuery } from '../../../../common/utils/apiUtils.js';
 
 Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
 const messages = Messages.loadMessages('sfdx-hardis', 'org');
@@ -91,9 +91,12 @@ export default class OrgPurgeFlow extends SfCommand<any> {
   // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
   public static requiresProject = false;
 
+  protected debugMode = false;
   protected statusFilter: string[] = [];
   protected nameFilter: string | null = null;
   protected username: string;
+  protected deleteFlowInterviews: boolean;
+  protected allowPurgeFailure: boolean;
   protected flowRecordsRaw: any[];
   protected flowRecords: any[];
   protected deletedRecords: any[] = [];
@@ -104,13 +107,13 @@ export default class OrgPurgeFlow extends SfCommand<any> {
     const { flags } = await this.parse(OrgPurgeFlow);
     const prompt = flags.prompt === false ? false : true;
     this.nameFilter = flags.name || null;
-    const allowPurgeFailure = flags.allowpurgefailure === false ? false : true;
-    // const deleteFlowInterviews = flags["delete-flow-interviews"] || false;
-    const debugMode = flags.debug || false;
+    this.allowPurgeFailure = flags.allowpurgefailure === false ? false : true;
+    this.deleteFlowInterviews = flags["delete-flow-interviews"] || false;
+    this.debugMode = flags.debug || false;
     this.username = flags['target-org'].getUsername();
 
     // List flows to delete, prompt user if not in CI and not send as arguments
-    const manageableConstraint = await this.getFlowsScope(flags, debugMode);
+    const manageableConstraint = await this.getFlowsScope(flags);
 
     // Check we don't delete active Flows
     if (this.statusFilter.includes('Active')) {
@@ -118,7 +121,7 @@ export default class OrgPurgeFlow extends SfCommand<any> {
     }
 
     // Build query with name filter if sent
-    await this.listFlowVersionsToDelete(manageableConstraint, debugMode);
+    await this.listFlowVersionsToDelete(manageableConstraint);
 
     // Check empty result
     if (this.flowRecordsRaw.length === 0) {
@@ -146,7 +149,7 @@ export default class OrgPurgeFlow extends SfCommand<any> {
 
     // Perform deletion
     const conn = flags['target-org'].getConnection();
-    await this.processDelete(conn, allowPurgeFailure);
+    await this.processDeleteFlowVersions(conn, true);
 
     const summary =
       this.deletedRecords.length > 0
@@ -157,8 +160,9 @@ export default class OrgPurgeFlow extends SfCommand<any> {
     return { orgId: flags['target-org'].getOrgId(), outputString: summary };
   }
 
-  private async processDelete(conn: any, allowPurgeFailure: boolean) {
-    const deleteResults = await bulkDeleteTooling('Flow', this.flowRecords, conn);
+  private async processDeleteFlowVersions(conn: any, tryDeleteInterviews: boolean) {
+    const recordsIds = this.flowRecords.map((record) => record.Id);
+    const deleteResults = await bulkDeleteTooling('Flow', recordsIds, conn);
     for (const deleteRes of deleteResults.results) {
       if (deleteRes.success) {
         this.deletedRecords.push(deleteRes);
@@ -167,9 +171,13 @@ export default class OrgPurgeFlow extends SfCommand<any> {
         this.deletedErrors.push(deleteRes);
       }
     }
+    if (this.deletedErrors.length > 0 && (this.deleteFlowInterviews === true || !isCI) && tryDeleteInterviews === true) {
+      await this.manageDeleteFlowInterviews(conn);
+    }
+
     if (this.deletedErrors.length > 0) {
       const errMsg = `[sfdx-hardis] There have been errors while deleting ${this.deletedErrors.length} record(s): \n${JSON.stringify(this.deletedErrors)}`;
-      if (allowPurgeFailure) {
+      if (this.allowPurgeFailure) {
         uxLog(this, c.yellow(errMsg));
       } else {
         throw new SfError(
@@ -179,6 +187,42 @@ export default class OrgPurgeFlow extends SfCommand<any> {
         );
       }
     }
+  }
+
+  private async manageDeleteFlowInterviews(conn: any) {
+    // Gather flow interviews that prevent deleting flow versions
+    const flowInterviewsIds: string[] = [];
+    this.flowRecords = [];
+    const extractInterviewsRegex = /Flow Interview - ([a-zA-Z0-9]{15}|[a-zA-Z0-9]{18})/gm;
+    for (const deletedError of this.deletedErrors) {
+      this.flowRecords.push({ Id: deletedError.Id });
+      const errorflowInterviewIds = await extractRegexMatches(extractInterviewsRegex, deletedError.error);
+      flowInterviewsIds.push(...[...new Set(errorflowInterviewIds)]); // make interview Ids unique
+    }
+    if (flowInterviewsIds.length === 0) {
+      return;
+    }
+    // Display flows & Prompt user if not in CI
+    await this.displayFlowInterviewToDelete(flowInterviewsIds, conn);
+    if (!isCI) {
+      const confirmDelete = await prompts({
+        type: 'confirm',
+        name: 'value',
+        message: c.cyanBright(`Do you confirm you want to delete ${flowInterviewsIds.length} Flow Interviews ?`),
+      });
+      if (confirmDelete.value === false) {
+        uxLog(this, c.magenta('Action cancelled by user'));
+        return { outputString: 'Action cancelled by user' };
+      }
+    }
+    // Delete flow interviews
+    const deleteInterviewResults = await bulkDelete('FlowInterview', flowInterviewsIds, conn);
+    this.deletedRecords.push(deleteInterviewResults?.successfulResults || [])
+    this.deletedErrors = deleteInterviewResults?.failedResults || [];
+    // Try to delete flow versions again
+    uxLog(this, c.cyan(`Trying again to delete flow versions after deleting flow interviews...`));
+    this.flowRecords = [...new Set(this.flowRecords)]; // Make list unique
+    await this.processDeleteFlowVersions(conn, false);
   }
 
   private formatFlowRecords() {
@@ -195,7 +239,7 @@ export default class OrgPurgeFlow extends SfCommand<any> {
     uxLog(this, `[sfdx-hardis] Found ${c.bold(this.flowRecords.length)} records:\n${c.yellow(columnify(this.flowRecords))}`);
   }
 
-  private async listFlowVersionsToDelete(manageableConstraint: string, debugMode: any) {
+  private async listFlowVersionsToDelete(manageableConstraint: string) {
     let query = `SELECT Id,MasterLabel,VersionNumber,Status,Description,Definition.DeveloperName FROM Flow WHERE ${manageableConstraint} AND Status IN ('${this.statusFilter.join(
       "','"
     )}')`;
@@ -207,13 +251,13 @@ export default class OrgPurgeFlow extends SfCommand<any> {
     const flowQueryCommand = 'sf data query ' + ` --query "${query}"` + ` --target-org ${this.username}` + ' --use-tooling-api';
     const flowQueryRes = await execSfdxJson(flowQueryCommand, this, {
       output: false,
-      debug: debugMode,
+      debug: this.debugMode,
       fail: true,
     });
     this.flowRecordsRaw = flowQueryRes?.result?.records || flowQueryRes.records || [];
   }
 
-  private async getFlowsScope(flags, debugMode: any) {
+  private async getFlowsScope(flags) {
     const manageableConstraint = "ManageableState IN ('deprecatedEditable','installedEditable','unmanaged')";
     if (flags.status) {
       // Input parameter used
@@ -229,7 +273,7 @@ export default class OrgPurgeFlow extends SfCommand<any> {
         ' --use-tooling-api';
       const allFlowQueryRes = await execSfdxJson(allFlowQueryCommand, this, {
         output: false,
-        debug: debugMode,
+        debug: this.debugMode,
         fail: true,
       });
       const flowRecordsRaw = allFlowQueryRes?.result?.records || allFlowQueryRes.records || [];
@@ -262,5 +306,13 @@ export default class OrgPurgeFlow extends SfCommand<any> {
       this.statusFilter = selectStatus.status;
     }
     return manageableConstraint;
+  }
+
+  private async displayFlowInterviewToDelete(flowVInterviewIds: string[], conn: any) {
+    const query = "SELECT Name,InterviewLabel,InterviewStatus,CreatedBy.Username,CreatedDate,LastModifiedDate " +
+      `FROM FlowInterview WHERE Id IN ('${flowVInterviewIds.join("','")}')` +
+      " ORDER BY Name";
+    const flowsInterviewsToDelete = (await bulkQuery(query, conn)).records;
+    uxLog(this, c.yellow(`Flow interviews to be deleted would be the following:\n${columnify(flowsInterviewsToDelete)}`));
   }
 }